feat: expose mock generator
